Fix compilation with FPC 2.4.4 --- needs Single() typecast to choose overloaded version.
git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle_game_engine@10775 afe8dcc5-9094-451b-8773-5f495042a737

+++ tests/testvectormath.pas
@@ -662,7 +662,7 @@ begin
 
   { rotate and move x = 0 plane to make z = 10 }
   DoTest(Vector4Single(1, 0, 0, 0),
-    TranslationMatrix(0, 0, 10) * RotationMatrixDeg(90, 0, 1, 0),
+    TranslationMatrix(Single(0), 0, 10) * RotationMatrixDeg(90, 0, 1, 0),
     [ Vector3Single( 10,  10, 10),
       Vector3Single(-10,  10, 10),
       Vector3Single( 10, -10, 10),
@@ -679,7 +679,7 @@ begin
   { rotate and move and scale x = 0 plane to make z = 100 }
   DoTest(Vector4Single(1, 0, 0, 0),
     ScalingMatrix(Vector3Single(10, 10, 10)) *
-    TranslationMatrix(0, 0, 10) *
+    TranslationMatrix(Single(0), 0, 10) *
     RotationMatrixDeg(90, 0, 1, 0),
     [ Vector3Single( 10,  10, 100),
       Vector3Single(-10,  10, 100),
